Django/requirements.txt
Django == 3.2.12
psycopg2-binary >= 2.8
djangorestframework == 3.13.1
django-cors-headers==3.13.0